Revert "chore: VR improvements (#3374)"
This reverts commit 0a60a3ede857f03e276978b9623642d40705d00a.
