+++ package.json
@@ -164,7 +164,7 @@
     {
       "path": "lib/components/internal/widget-exports.js",
       "brotli": false,
-      "limit": "754 kB",
+      "limit": "753 kB",
       "ignore": "react-dom"
     }
   ],
