+++ src/code-editor/styles.scss
@@ -13,8 +13,8 @@
 .code-editor {
   @include styles.styles-reset;
   display: block;
-  border-block: awsui.$border-width-field solid awsui.$color-border-code-editor-default;
-  border-inline: awsui.$border-width-field solid awsui.$color-border-code-editor-default;
+  border-block: awsui.$border-item-width solid awsui.$color-border-code-editor-default;
+  border-inline: awsui.$border-item-width solid awsui.$color-border-code-editor-default;
   border-start-start-radius: awsui.$border-radius-code-editor;
   border-start-end-radius: awsui.$border-radius-code-editor;
   border-end-start-radius: awsui.$border-radius-code-editor;
@@ -46,7 +46,7 @@
   container-type: inline-size;
   display: flex;
   vertical-align: middle;
-  border-block-start: awsui.$border-width-field solid awsui.$color-border-code-editor-default;
+  border-block-start: awsui.$border-item-width solid awsui.$color-border-code-editor-default;
   background-color: awsui.$color-background-code-editor-status-bar;
   @include styles.text-wrapping;
 
@@ -60,7 +60,7 @@
     display: flex;
     flex-wrap: wrap;
     padding-inline-start: awsui.$space-l;
-    border-inline-end: awsui.$border-width-field solid awsui.$color-border-code-editor-default;
+    border-inline-end: awsui.$border-item-width solid awsui.$color-border-code-editor-default;
   }
 
   &__right {
