+++ src/internal/components/button-trigger/styles.scss
@@ -29,16 +29,14 @@ $padding-block-inner-filtering-token: 0px;
   border-start-end-radius: styles.$control-border-radius;
   border-end-start-radius: styles.$control-border-radius;
   border-end-end-radius: styles.$control-border-radius;
-  border-block: awsui.$border-width-field solid awsui.$color-border-input-default;
-  border-inline: awsui.$border-width-field solid awsui.$color-border-input-default;
+  border-block: styles.$control-border-width solid awsui.$color-border-input-default;
+  border-inline: styles.$control-border-width solid awsui.$color-border-input-default;
   min-block-size: awsui.$size-vertical-input;
 
   &.in-filtering-token-root,
   &.in-filtering-token-nested {
     padding-block: $padding-block-inner-filtering-token;
     padding-inline: $padding-inline-inner-filtering-token;
-    border-block-width: awsui.$border-width-token;
-    border-inline-width: awsui.$border-width-token;
 
     border-color: tokenGroup.$token-border-color;
     border-start-end-radius: 0;
