+++ src/internal/components/token-list/styles.scss
@@ -75,8 +75,8 @@
   display: flex;
   align-items: center;
   background-color: transparent;
-  border-block: awsui.$border-width-button solid transparent;
-  border-inline: awsui.$border-width-button solid transparent;
+  border-block: styles.$control-border-width solid transparent;
+  border-inline: styles.$control-border-width solid transparent;
 
   padding-block: 0;
   padding-inline: 0;
