+++ src/internal/styles/forms/mixins.scss
@@ -98,8 +98,8 @@
   // Using a special transparent outline only visible in Windows High Contrast Mode.
   // See focus-highlight above.
   outline: 2px dotted transparent;
-  border-block: awsui.$border-width-button solid awsui.$color-border-item-focused;
-  border-inline: awsui.$border-width-button solid awsui.$color-border-item-focused;
+  border-block: constants.$control-border-width solid awsui.$color-border-item-focused;
+  border-inline: constants.$control-border-width solid awsui.$color-border-item-focused;
   border-start-start-radius: $border-radius;
   border-start-end-radius: $border-radius;
   border-end-start-radius: $border-radius;
@@ -114,8 +114,8 @@
   // See focus-highlight above.
   outline: 2px dotted transparent;
 
-  border-block: awsui.$border-width-field solid awsui.$color-border-input-focused;
-  border-inline: awsui.$border-width-field solid awsui.$color-border-input-focused;
+  border-block: constants.$control-border-width solid awsui.$color-border-input-focused;
+  border-inline: constants.$control-border-width solid awsui.$color-border-input-focused;
   border-start-start-radius: $border-radius;
   border-start-end-radius: $border-radius;
   border-end-start-radius: $border-radius;
@@ -126,16 +126,16 @@
 
 @mixin form-disabled-element {
   background-color: awsui.$color-background-input-disabled;
-  border-block: awsui.$border-width-field solid awsui.$color-border-input-disabled;
-  border-inline: awsui.$border-width-field solid awsui.$color-border-input-disabled;
+  border-block: constants.$control-border-width solid awsui.$color-border-input-disabled;
+  border-inline: constants.$control-border-width solid awsui.$color-border-input-disabled;
   color: awsui.$color-text-input-disabled;
   cursor: auto;
 }
 
 @mixin form-readonly-element {
   background-color: awsui.$color-background-input-default;
-  border-block: awsui.$border-width-field solid awsui.$color-border-input-disabled;
-  border-inline: awsui.$border-width-field solid awsui.$color-border-input-disabled;
+  border-block: constants.$control-border-width solid awsui.$color-border-input-disabled;
+  border-inline: constants.$control-border-width solid awsui.$color-border-input-disabled;
 }
 
 @mixin form-placeholder {
@@ -238,8 +238,8 @@
     inset-inline-start: 0px;
     inset-inline-end: 0px;
     background-color: transparent;
-    border-block: awsui.$border-width-popover solid awsui.$color-border-dropdown-container;
-    border-inline: awsui.$border-width-popover solid awsui.$color-border-dropdown-container;
+    border-block: awsui.$border-width-field solid awsui.$color-border-dropdown-container;
+    border-inline: awsui.$border-width-field solid awsui.$color-border-dropdown-container;
     border-start-start-radius: awsui.$border-radius-dropdown;
     border-start-end-radius: awsui.$border-radius-dropdown;
     border-end-start-radius: awsui.$border-radius-dropdown;
