+++ src/pagination/styles.scss
@@ -23,8 +23,8 @@
 .button,
 .dots {
   min-inline-size: awsui.$space-l;
-  border-block: awsui.$border-width-button solid transparent;
-  border-inline: awsui.$border-width-button solid transparent;
+  border-block: styles.$control-border-width solid transparent;
+  border-inline: styles.$control-border-width solid transparent;
 }
 
 .button {
