+++ src/property-filter/styles.scss
@@ -213,7 +213,7 @@ $operator-field-width: 120px;
 
 .results {
   // Align with the filter input's vertical padding and border sizes.
-  padding-block: calc(#{styles.$control-padding-vertical} + #{awsui.$border-width-field});
+  padding-block: calc(#{styles.$control-padding-vertical} + #{styles.$control-border-width});
   padding-inline: 0;
 }
