+++ src/segmented-control/styles.scss
@@ -33,8 +33,7 @@
   border-end-end-radius: awsui.$border-radius-button;
   align-items: center;
   min-block-size: calc(awsui.$line-height-body-m + awsui.$space-static-xxs);
-  padding-block: calc(seg.$segment-gap / 2 - awsui.$border-width-field);
-  padding-inline: calc(seg.$segment-gap - awsui.$border-width-field);
+  padding-inline: awsui.$space-static-xxxs;
   gap: calc(#{seg.$segment-gap} * 2 + #{seg.$segment-divider-width});
 
   &:not(.refresh) {
