+++ src/table/body-cell/styles.scss
@@ -213,9 +213,6 @@ $cell-negative-space-vertical: 2px;
   &-shaded {
     background: awsui.$color-background-cell-shaded;
   }
-  &.has-striped-rows:not(.body-cell-selected):not(.body-cell-last-row) {
-    border-block-end-color: awsui.$color-border-cell-shaded;
-  }
   &-selected {
     background-color: awsui.$color-background-item-selected;
     border-block-start: $selected-border;
