+++ style-dictionary/utils/token-names.ts
@@ -19,7 +19,6 @@ export type ColorPaletteTokenName =
   | 'colorGreen600'
   | 'colorGreen700'
   | 'colorGreen900'
-  | 'colorGrey50'
   | 'colorGrey100'
   | 'colorGrey125'
   | 'colorGrey150'
@@ -338,7 +337,6 @@ export type ColorsTokenName =
   | 'colorTextButtonPrimaryDisabled'
   | 'colorBorderCalendarGrid'
   | 'colorBorderCalendarGridSelectedFocusRing'
-  | 'colorBorderCellShaded'
   | 'colorBorderCodeEditorAceActiveLineLightTheme'
   | 'colorBorderCodeEditorAceActiveLineDarkTheme'
   | 'colorBorderCodeEditorDefault'
