+++ style-dictionary/visual-refresh/colors.ts
@@ -118,7 +118,6 @@ const tokens: StyleDictionary.ColorsDictionary = {
   colorTextButtonPrimaryDisabled: { light: '{colorGrey500}', dark: '{colorGrey500}' },
   colorBorderCalendarGrid: 'transparent',
   colorBorderCalendarGridSelectedFocusRing: { light: '{colorGrey100}', dark: '{colorGrey800}' },
-  colorBorderCellShaded: { light: '{colorGrey300}', dark: '{colorGrey650}' },
   colorBorderCodeEditorAceActiveLineLightTheme: '{colorGrey300}',
   colorBorderCodeEditorAceActiveLineDarkTheme: '{colorGrey550}',
   colorBorderCodeEditorDefault: { light: '{colorGrey300}', dark: '{colorGrey550}' },
@@ -143,12 +142,12 @@ const tokens: StyleDictionary.ColorsDictionary = {
   colorBorderEditableCellHover: '{colorBorderDropdownItemHover}',
   colorBorderInputDefault: { light: '{colorGrey500}', dark: '{colorGrey550}' },
   colorBorderInputDisabled: '{colorBackgroundInputDisabled}',
-  colorBorderInputFocused: { light: '{colorBlue600}', dark: '{colorBlue500}' },
+  colorBorderInputFocused: { light: '{colorBlue800}', dark: '{colorBlue300}' },
   colorBorderItemFocused: { light: '{colorBlue600}', dark: '{colorBlue500}' },
   colorBorderDropdownItemFocused: { light: '{colorGrey600}', dark: '{colorGrey300}' },
   colorBorderItemPlaceholder: '{colorBorderItemSelected}',
   colorBorderItemSelected: { light: '{colorBlue600}', dark: '{colorBlue500}' },
-  colorBorderLayout: { light: '{colorGrey350}', dark: '{colorGrey600}' },
+  colorBorderLayout: { light: '{colorGrey400}', dark: '{colorGrey600}' },
   colorBorderNotificationStackBar: '{colorGrey700}',
   colorBorderPanelHeader: '{colorBorderDividerDefault}',
   colorBorderPopover: { light: '{colorGrey400}', dark: '{colorGrey550}' },
