--- style-dictionary/visual-refresh/contexts/app-layout-toolbar.ts
@@ -1,14 +1,0 @@
-// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
-// SPDX-License-Identifier: Apache-2.0
-import merge from 'lodash/merge.js';
-
-import { expandColorDictionary } from '../../utils/index.js';
-import { StyleDictionary } from '../../utils/interfaces.js';
-
-const tokens = {
-  colorBackgroundLayoutMain: { light: '{colorGrey50}', dark: '{colorGrey850}' },
-};
-
-const expandedTokens: StyleDictionary.ExpandedColorScopeDictionary = expandColorDictionary(merge({}, { ...tokens }));
-
-export { expandedTokens as tokens };
